Move startup modal to page
