+++ src/common/PageRoute.ts
@@ -1,7 +1,7 @@
 export const enum PageRoute {
   ANIMATOR = "/animator",
   EXPORT_VIDEO_MODAL = "/animator/exportVideoModal",
-  STARTUP_MODAL = "/startupModal",
+  STARTUP_PAGE = "/startupPage",
   PROJECT_SETTINGS_MODAL = "/animator/projectSettingsModal",
   PREFERENCES_MODAL = "/animator/preferencesModal",
 }
