+++ src/renderer/components/startup/NewsFeed/NewsFeed.css
