+++ src/renderer/components/startup/NewsFeed/NewsFeed.tsx
